chore(render): split markdown renderer into modules
Keep the markdown renderer easier to maintain by separating each output section into focused modules without changing the public rendering API.
